[fix] allow all cors temporary
(cherry picked from commit 2de81b6e53025e51d4c09edb82a1393de7c090ec)

+++ src/main/java/com/example/gistcompetitioncnserver/user/UserService.java
@@ -57,7 +57,11 @@ public class UserService implements UserDetailsService {
         if (userExists){
             Optional<EmailConfirmationToken> userEmailToken = emailConfirmationTokenRepository.findByUserId(userInfo.get().getId()) ;
 
-            if((userInfo.get().getEmail()).equals((user.getEmail()))){
+            if (!userInfo.get().isEnabled()){
+                throw new IllegalStateException("이미 존재하는 이메일입니다.");
+            }
+
+            if((userInfo.get().getEmail()).equals((user.getEmail())) && !userInfo.get().isEnabled()){
                 emailConfirmationTokenService.deleteToken(userEmailToken.get().getToken());
                 String token = UUID.randomUUID().toString();
